Add: Sample Project implementation
