+++ sample/build.gradle
@@ -12,6 +12,8 @@ buildscript {
     }
 }
 
+
+
 repositories {
     mavenCentral()
 }
@@ -19,4 +21,5 @@ repositories {
 dependencies {
     compile project(":mockserver")
     provided 'javax.annotation:jsr250-api:1.0'
+
 }
