--- sample/src/main/java/com/byoutline/mockserver/sample/config.json
@@ -1,10 +1,0 @@
-{
-  "port": 8099,
-  "requests": [
-    {
-      "method": "GET",
-      "path": "/books",
-      "response file": "books.json"
-    }
-  ]
-}
