+++ sample/src/main/resources/config.json
@@ -4,7 +4,7 @@
     {
       "method": "GET",
       "path": "/books",
-      "response file": "books.json"
+      "response file": "books/books.json"
     }
   ]
 }
