--- sample/src/main/resources/static/books/books.json
@@ -1,17 +1,0 @@
-{
-  "title": "1000 miles",
-  "properties": {
-    "firstName": {
-      "type": "1"
-    },
-    "lastName": {
-      "type": "string"
-    },
-    "age": {
-      "description": "Age in years",
-      "type": "integer",
-      "minimum": 0
-    }
-  },
-  "required": ["firstName", "lastName"]
-}
